refactoring wardrobe to proper mvc
