+++ SmartFit/WardrobeView.swift
@@ -69,6 +69,50 @@ struct WardrobeView: View {
     }
 }
 
+struct ItemCard: View {
+    let item: WardrobeItem
+
+    var body: some View {
+        VStack(alignment: .leading) {
+            if let imageData = item.image_data,
+               let base64 = imageData.components(separatedBy: ",").last,
+               let data = Data(base64Encoded: base64),
+               let uiImage = UIImage(data: data) {
+                Image(uiImage: uiImage)
+                    .resizable()
+                    .aspectRatio(contentMode: .fill)
+                    .frame(height: 150)
+                    .clipped()
+                    .cornerRadius(8)
+            } else {
+                Rectangle()
+                    .fill(Color.gray.opacity(0.2))
+                    .frame(height: 150)
+                    .cornerRadius(8)
+                    .overlay(
+                        Image(systemName: "tshirt")
+                            .font(.system(size: 40))
+                            .foregroundColor(.gray)
+                    )
+            }
+
+            Text(item.name)
+                .font(.headline)
+                .lineLimit(1)
+
+            if let brand = item.brand {
+                Text(brand)
+                    .font(.caption)
+                    .foregroundColor(.secondary)
+            }
+        }
+        .padding(8)
+        .background(Color.white)
+        .cornerRadius(12)
+        .shadow(radius: 2)
+    }
+}
+
 struct AddItemSheet: View {
     @Environment(\.dismiss) var dismiss
     @ObservedObject var controller: WardrobeController
